fix: resolve build errors for cross-platform compilation
- Fix SPSC queue self-include causing infinite recursion
- Add static_cast for timing.hpp sign conversion warning
- Add missing #include <functional> in event_journal.hpp
- Enable exceptions for test compatibility
- Fix snprintf buffer size warning in main.cpp

Build now passes locally. Simulation runs at 3.3M orders/sec.

+++ include/utils/timing.hpp
@@ -21,7 +21,7 @@ public:
 #else
         // Fallback for non-x86
         auto now = std::chrono::high_resolution_clock::now();
-        return now.time_since_epoch().count();
+        return static_cast<uint64_t>(now.time_since_epoch().count());
 #endif
     }
